🧹 code health: simplify SettingsPage by extracting ProfileForm
