+++ .Jules/pre-commit-instructions.md
@@ -1,0 +1,7 @@
+# Pre-Commit Instructions
+
+Verify everything works locally before pushing.
+
+1.  Run the tests to ensure functionality isn't broken.
+2.  Run the linter and code formatter to ensure formatting isn't broken.
+3.  Check if any new documentation is needed, or if an architectural decision needs to be recorded.
