+++ get_file.js
@@ -1,0 +1,3 @@
+const fs = require('fs');
+const content = fs.readFileSync('apps/web/src/app/settings/page.tsx', 'utf-8');
+console.log(content);
