+++ read_file_script.py
@@ -1,0 +1,3 @@
+import sys
+with open('apps/web/src/app/settings/page.tsx', 'r') as f:
+    print(f.read())
